+++ services/GateWay/src/core/connection/connection.cc
@@ -8,11 +8,11 @@
 #include <functional>
 #include <global/Global.hpp>
 #include <string>
+#include <tools/Defer.hpp>
 #include <tools/Logger.hpp>
 #include <utility>
 #include <utils/common/type.hpp>
 #include <utils/url/url.hpp>
-#include <tools/Defer.hpp>
 
 namespace core
 {